[geometry] move /odometry headers to /geometry
