+++ ecl_geometry/include/ecl/CMakeLists.txt
@@ -3,7 +3,6 @@
 ###############################################################################
 
 add_subdirectory(geometry)
-add_subdirectory(odometry)
 
 ###############################################################################
 # FILES
