+++ ecl_geometry/include/ecl/geometry/odometry_helper.hpp
