+++ ecl_geometry/include/ecl/geometry/odometry_typedefs.hpp
